requirements.txt
alembic==1.17.2
asn1crypto==1.5.1
bidict==0.23.1
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
gevent==25.9.1
gevent-websocket==0.10.1
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pg8000==1.31.5
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.3
python-socketio==5.15.0
scramp==1.4.6
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
urllib3==2.6.0
Werkzeug==3.1.3
wsproto==1.3.2
zope.event==6.1
zope.interface==8.1.1
cryptography
requests